modified the leads team memebers analytics blocks and analytice page functinalities

+++ src/main/java/com/istlgroup/istl_group_crm_backend/wrapperClasses/DashboardDTO.java
@@ -153,7 +153,7 @@ public class DashboardDTO {
         private BigDecimal  revenue;
         private long        followupsDone;
         private long        followupsPending;
-        private int         conversionRate;
+        private double      conversionRate;
     }
 
     @Data @Builder @NoArgsConstructor @AllArgsConstructor
